multiTable.go

+++ server/model/basicdata/request/multiTable.go
@@ -8,6 +8,6 @@ package request
 
 // 接收 教学班id 和学生id 的结构体
 type StuTeachClass struct {
-	Teach_class_id int64   `json:"teach_class_id"`
-	Student_ids    []int64 `json:"student_ids"`
+	Teach_class_id int   `json:"teach_class_id"`
+	Student_ids    []int `json:"student_ids"`
 }
